+++ Modules/vtkFMRIEngine/cxx/vtkParzenDensityEstimation.cxx
@@ -213,7 +213,7 @@ void vtkParzenDensityEstimation::SimpleExecute(vtkImageData *input, vtkImageData
     outputArray->InsertNextValue(0.0);
      
   // calculation of probability density     
-  for (unsigned long int j=0; j<numTrData; j++){
+  for (unsigned long int j=0; j<(unsigned long int)numTrData; j++){
     for (unsigned long int n=0; n<size; n++){    
       power = pow((float)((trainingUse->GetValue(j))-(inputArray->GetValue(n))),2);
       outputArray->SetValue(n, (outputArray->GetValue(n)) + 1/(numTrData*pow(2.0*pi,0.5)*hBest)*exp(-0.5*power/pow((float)hBest,2)));
